+++ src/lib/storage/index.ts
@@ -1,0 +1,6 @@
+export const STORAGE_KEYS = {
+  REFRESH_TOKEN: "auth.refresh-token",
+  TOKEN: "auth.token",
+  SYNC_WORK_DELAY_DATE: "user.okami-sync-work-delay-date",
+  LAST_WORK_CLICKED: "user.okami-last-work-clicked",
+} as const;
